pyproject.toml
[tool.poetry]
name = "aguefort"
version = "0.0.1"
description = "A chatbot for asking about D&D tips powered by Dropout's Adventuring Academy knowledge and LLMs."
authors = ["Allan Reyes <allan_bassail@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
yaspin = "^3.0.1"
langchain-text-splitters = "^0.0.1"
langchain-community = "^0.0.29"
faiss-cpu = "^1.8.0"
sentence-transformers = "^2.5.1"
gradio = "^4.21.0"
langchain = "^0.1.13"
boto3 = "^1.34.68"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
autohooks = "^24.2.0"
autohooks-plugin-pylint = "^23.10.0"
autohooks-plugin-black = "^23.10.0"
autohooks-plugin-isort = "^23.10.0"

[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.isort", "autohooks.plugins.pylint"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.pylint.format]
max-line-length=120

[tool.pylint.messages_control]
disable=["fixme"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
